+++ connector/src/test/scala/com/vertica/spark/datasource/wrappers/json/VerticaJsonTableWrapperTest.scala
@@ -1,5 +1,18 @@
-package com.vertica.spark.datasource.wrappers
+// (c) Copyright [2020-2021] Micro Focus or one of its affiliates.
+// Licensed under the Apache License, Version 2.0 (the "License");
+// You may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+package com.vertica.spark.datasource.wrappers.json
 
+import com.vertica.spark.datasource.wrappers.VerticaScanWrapperBuilder
 import org.apache.spark.sql.SparkSession
 import org.apache.spark.sql.execution.datasources.FileFormat
 import org.apache.spark.sql.execution.datasources.json.JsonFileFormat
